Clean-up of client code
